% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_recessions.R
\docType{data}
\name{recessions}
\alias{recessions}
\alias{fetch_recessions}
\title{Recessions}
\format{
A data frame with the following variables: \itemize{ \item
\code{start_char, end_char}: Easily readable labels for the beginning
and end of the recession. \item \code{start_date, end_date}: Recession Dates
expressed in R datetime format, using the first day of the specified month.
\item \code{ongoing}: Logical. Whether or not the recession is ongoing as of
the latest available NBER data. }
}
\source{
\url{https://www.nber.org/data/cycles/cycle dates pasted.csv}
}
\usage{
recessions

fetch_recessions(url = NULL, quietly = FALSE)
}
\arguments{
\item{url}{Char, the web location of the NBER machine-readable CSV file. The
default, \code{NULL}, uses the most recently identified URL known to the
package development team.}

\item{quietly}{Logical, suppresses messages.}
}
\description{
The \code{recessions} table is published by NBER. The \code{fetch_recessions}
function downloads the current table from NBER and formats it for use by
other functions in this package.
}
\section{Functions}{
\itemize{
\item \code{fetch_recessions}: Function for fetching recessions from NBER website
}}

\examples{
recessions <- fetch_recessions()

\dontrun{
  # package maintainers run:
  recessions <- fetch_recessions()
  usethis::use_data(recessions, overwrite = TRUE)
}

}
\keyword{datasets}
